+++ apps/web/components/pages/auth/login.tsx
@@ -1,0 +1,268 @@
+"use client";
+
+import { zodResolver } from "@hookform/resolvers/zod";
+import classNames from "classnames";
+import { signIn } from "next-auth/react";
+import Link from "next/link";
+import { useRouter } from "next/navigation";
+import { useEffect, useState } from "react";
+import { FormProvider, useForm } from "react-hook-form";
+import { FaGoogle } from "react-icons/fa";
+import { z } from "zod";
+
+import { SAMLLogin } from "@calcom/features/auth/SAMLLogin";
+import { ErrorCode } from "@calcom/features/auth/lib/ErrorCode";
+import { HOSTED_CAL_FEATURES, WEBAPP_URL, WEBSITE_URL } from "@calcom/lib/constants";
+import { getSafeRedirectUrl } from "@calcom/lib/getSafeRedirectUrl";
+import { useCompatSearchParams } from "@calcom/lib/hooks/useCompatSearchParams";
+import { useLocale } from "@calcom/lib/hooks/useLocale";
+import { collectPageParameters, telemetryEventTypes, useTelemetry } from "@calcom/lib/telemetry";
+import { trpc } from "@calcom/trpc/react";
+import { Alert, Button, EmailField, PasswordField } from "@calcom/ui";
+import { ArrowLeft, Lock } from "@calcom/ui/components/icon";
+
+import AddToHomescreen from "@components/AddToHomescreen";
+import BackupCode from "@components/auth/BackupCode";
+import TwoFactor from "@components/auth/TwoFactor";
+import AuthContainer from "@components/ui/AuthContainer";
+
+import { type PageProps } from "@server/lib/auth/login/getServerSideProps";
+
+interface LoginValues {
+  email: string;
+  password: string;
+  totpCode: string;
+  backupCode: string;
+  csrfToken: string;
+}
+function Login({
+  csrfToken,
+  isGoogleLoginEnabled,
+  isSAMLLoginEnabled,
+  samlTenantID,
+  samlProductID,
+  totpEmail,
+}: // eslint-disable-next-line @typescript-eslint/ban-types
+PageProps) {
+  const searchParams = useCompatSearchParams();
+  const { t } = useLocale();
+  const router = useRouter();
+  const formSchema = z
+    .object({
+      email: z
+        .string()
+        .min(1, `${t("error_required_field")}`)
+        .email(`${t("enter_valid_email")}`),
+      password: !!totpEmail ? z.literal("") : z.string().min(1, `${t("error_required_field")}`),
+    })
+    // Passthrough other fields like totpCode
+    .passthrough();
+  const methods = useForm<LoginValues>({ resolver: zodResolver(formSchema) });
+  const { register, formState } = methods;
+  const [twoFactorRequired, setTwoFactorRequired] = useState(!!totpEmail || false);
+  const [twoFactorLostAccess, setTwoFactorLostAccess] = useState(false);
+  const [errorMessage, setErrorMessage] = useState<string | null>(null);
+
+  const errorMessages: { [key: string]: string } = {
+    // [ErrorCode.SecondFactorRequired]: t("2fa_enabled_instructions"),
+    // Don't leak information about whether an email is registered or not
+    [ErrorCode.IncorrectEmailPassword]: t("incorrect_email_password"),
+    [ErrorCode.IncorrectTwoFactorCode]: `${t("incorrect_2fa_code")} ${t("please_try_again")}`,
+    [ErrorCode.InternalServerError]: `${t("something_went_wrong")} ${t("please_try_again_and_contact_us")}`,
+    [ErrorCode.ThirdPartyIdentityProviderEnabled]: t("account_created_with_identity_provider"),
+  };
+
+  const telemetry = useTelemetry();
+
+  let callbackUrl = searchParams?.get("callbackUrl") || "";
+
+  if (/"\//.test(callbackUrl)) callbackUrl = callbackUrl.substring(1);
+
+  // If not absolute URL, make it absolute
+  if (!/^https?:\/\//.test(callbackUrl)) {
+    callbackUrl = `${WEBAPP_URL}/${callbackUrl}`;
+  }
+
+  const safeCallbackUrl = getSafeRedirectUrl(callbackUrl);
+
+  callbackUrl = safeCallbackUrl || "";
+
+  const LoginFooter = (
+    <Link href={`${WEBSITE_URL}/signup`} className="text-brand-500 font-medium">
+      {t("dont_have_an_account")}
+    </Link>
+  );
+
+  const TwoFactorFooter = (
+    <>
+      <Button
+        onClick={() => {
+          if (twoFactorLostAccess) {
+            setTwoFactorLostAccess(false);
+            methods.setValue("backupCode", "");
+          } else {
+            setTwoFactorRequired(false);
+            methods.setValue("totpCode", "");
+          }
+          setErrorMessage(null);
+        }}
+        StartIcon={ArrowLeft}
+        color="minimal">
+        {t("go_back")}
+      </Button>
+      {!twoFactorLostAccess ? (
+        <Button
+          onClick={() => {
+            setTwoFactorLostAccess(true);
+            setErrorMessage(null);
+            methods.setValue("totpCode", "");
+          }}
+          StartIcon={Lock}
+          color="minimal">
+          {t("lost_access")}
+        </Button>
+      ) : null}
+    </>
+  );
+
+  const ExternalTotpFooter = (
+    <Button
+      onClick={() => {
+        window.location.replace("/");
+      }}
+      color="minimal">
+      {t("cancel")}
+    </Button>
+  );
+
+  const onSubmit = async (values: LoginValues) => {
+    setErrorMessage(null);
+    telemetry.event(telemetryEventTypes.login, collectPageParameters());
+    const res = await signIn<"credentials">("credentials", {
+      ...values,
+      callbackUrl,
+      redirect: false,
+    });
+    if (!res) setErrorMessage(errorMessages[ErrorCode.InternalServerError]);
+    // we're logged in! let's do a hard refresh to the desired url
+    else if (!res.error) router.push(callbackUrl);
+    else if (res.error === ErrorCode.SecondFactorRequired) setTwoFactorRequired(true);
+    else if (res.error === ErrorCode.IncorrectBackupCode) setErrorMessage(t("incorrect_backup_code"));
+    else if (res.error === ErrorCode.MissingBackupCodes) setErrorMessage(t("missing_backup_codes"));
+    // fallback if error not found
+    else setErrorMessage(errorMessages[res.error] || t("something_went_wrong"));
+  };
+
+  const { data, isPending, error } = trpc.viewer.public.ssoConnections.useQuery();
+
+  useEffect(
+    function refactorMeWithoutEffect() {
+      if (error) {
+        setErrorMessage(error.message);
+      }
+    },
+    [error]
+  );
+
+  const displaySSOLogin = HOSTED_CAL_FEATURES
+    ? true
+    : isSAMLLoginEnabled && !isPending && data?.connectionExists;
+
+  return (
+    <div className="dark:bg-brand dark:text-brand-contrast text-emphasis min-h-screen [--cal-brand-emphasis:#101010] [--cal-brand-subtle:#9CA3AF] [--cal-brand-text:white] [--cal-brand:#111827] dark:[--cal-brand-emphasis:#e1e1e1] dark:[--cal-brand-text:black] dark:[--cal-brand:white]">
+      <AuthContainer
+        title={t("login")}
+        description={t("login")}
+        showLogo
+        heading={twoFactorRequired ? t("2fa_code") : t("welcome_back")}
+        footerText={
+          twoFactorRequired
+            ? !totpEmail
+              ? TwoFactorFooter
+              : ExternalTotpFooter
+            : process.env.NEXT_PUBLIC_DISABLE_SIGNUP !== "true"
+            ? LoginFooter
+            : null
+        }>
+        <FormProvider {...methods}>
+          <form onSubmit={methods.handleSubmit(onSubmit)} noValidate data-testid="login-form">
+            <div>
+              <input defaultValue={csrfToken || undefined} type="hidden" hidden {...register("csrfToken")} />
+            </div>
+            <div className="space-y-6">
+              <div className={classNames("space-y-6", { hidden: twoFactorRequired })}>
+                <EmailField
+                  id="email"
+                  label={t("email_address")}
+                  defaultValue={totpEmail || (searchParams?.get("email") as string)}
+                  placeholder="john.doe@example.com"
+                  required
+                  {...register("email")}
+                />
+                <div className="relative">
+                  <PasswordField
+                    id="password"
+                    autoComplete="off"
+                    required={!totpEmail}
+                    className="mb-0"
+                    {...register("password")}
+                  />
+                  <div className="absolute -top-[2px] ltr:right-0 rtl:left-0">
+                    <Link
+                      href="/auth/forgot-password"
+                      tabIndex={-1}
+                      className="text-default text-sm font-medium">
+                      {t("forgot")}
+                    </Link>
+                  </div>
+                </div>
+              </div>
+
+              {twoFactorRequired ? !twoFactorLostAccess ? <TwoFactor center /> : <BackupCode center /> : null}
+
+              {errorMessage && <Alert severity="error" title={errorMessage} />}
+              <Button
+                type="submit"
+                color="primary"
+                disabled={formState.isSubmitting}
+                className="w-full justify-center">
+                {twoFactorRequired ? t("submit") : t("sign_in")}
+              </Button>
+            </div>
+          </form>
+          {!twoFactorRequired && (
+            <>
+              {(isGoogleLoginEnabled || displaySSOLogin) && <hr className="border-subtle my-8" />}
+              <div className="space-y-3">
+                {isGoogleLoginEnabled && (
+                  <Button
+                    color="secondary"
+                    className="w-full justify-center"
+                    disabled={formState.isSubmitting}
+                    data-testid="google"
+                    StartIcon={FaGoogle}
+                    onClick={async (e) => {
+                      e.preventDefault();
+                      await signIn("google");
+                    }}>
+                    {t("signin_with_google")}
+                  </Button>
+                )}
+                {displaySSOLogin && (
+                  <SAMLLogin
+                    samlTenantID={samlTenantID}
+                    samlProductID={samlProductID}
+                    setErrorMessage={setErrorMessage}
+                  />
+                )}
+              </div>
+            </>
+          )}
+        </FormProvider>
+      </AuthContainer>
+      <AddToHomescreen />
+    </div>
+  );
+}
+
+export default Login;
