+++ apps/web/pages/auth/error.tsx
@@ -1,60 +1,7 @@
-import type { GetStaticPropsContext } from "next";
-import Link from "next/link";
-import z from "zod";
-
-import { useCompatSearchParams } from "@calcom/lib/hooks/useCompatSearchParams";
-import { useLocale } from "@calcom/lib/hooks/useLocale";
-import { Button } from "@calcom/ui";
-import { X } from "@calcom/ui/components/icon";
-
 import PageWrapper from "@components/PageWrapper";
-import AuthContainer from "@components/ui/AuthContainer";
+import Error from "@components/pages/auth/error";
 
-import { getTranslations } from "@server/lib/getTranslations";
+export { getStaticProps } from "@lib/auth/error/getStaticProps";
 
-const querySchema = z.object({
-  error: z.string().optional(),
-});
-
-export default function Error() {
-  const { t } = useLocale();
-  const searchParams = useCompatSearchParams();
-  const { error } = querySchema.parse(searchParams);
-  const isTokenVerificationError = error?.toLowerCase() === "verification";
-  const errorMsg = isTokenVerificationError ? t("token_invalid_expired") : t("error_during_login");
-
-  return (
-    <AuthContainer title="" description="">
-      <div>
-        <div className="bg-error mx-auto flex h-12 w-12 items-center justify-center rounded-full">
-          <X className="h-6 w-6 text-red-600" />
-        </div>
-        <div className="mt-3 text-center sm:mt-5">
-          <h3 className="text-emphasis text-lg font-medium leading-6" id="modal-title">
-            {error}
-          </h3>
-          <div className="mt-2">
-            <p className="text-subtle text-sm">{errorMsg}</p>
-          </div>
-        </div>
-      </div>
-      <div className="mt-5 sm:mt-6">
-        <Link href="/auth/login" passHref legacyBehavior>
-          <Button className="flex w-full justify-center">{t("go_back_login")}</Button>
-        </Link>
-      </div>
-    </AuthContainer>
-  );
-}
-
+export default Error;
 Error.PageWrapper = PageWrapper;
-
-export const getStaticProps = async (context: GetStaticPropsContext) => {
-  const i18n = await getTranslations(context);
-
-  return {
-    props: {
-      i18n,
-    },
-  };
-};
